Add hint for Debian Bookworm ZFS Root Networking
Note that you may need to have appropriate firmware and networking
utilities installed to have a working network upon reboot. This is
varied so don't be overly specific but do point it out.

Adding this as I was caught out because my wifi adapter needed this on
my most recent installation.

Signed-off-by: Ben Walton <bdwalton@gmail.com>